requirements.txt
requests
flask
python-dateutil
pandas
pynamodb
pynamodb_attributes
tqdm
